requirements.txt
boto3==1.20.24
fastapi==0.115.5
mangum==0.19.0
python-multipart==0.0.5
redis==3.5.3
requests==2.22.0
rsa==4.7.2
starlette==0.41.3
uvicorn==0.32.1
pydantic==2.10.1
pydantic-core==2.27.1
PyJWT==2.3.0
email-validator==2.2.0
httpx~=0.27.2
pillow==10.4.0
pytest==6.2.5
babel==2.16.0
pycountry==24.6.1
annotated-types>=0.4.0
typing-extensions>=4.6.1
